+++ R/Bacon.R
@@ -6,7 +6,7 @@
 #' @docType package
 #' @author Maarten Blaauw <maarten.blaauw@qub.ac.uk> J. Andres Christen <jac@cimat.mx> 
 #' @importFrom grDevices dev.cur dev.off pdf dev.copy2pdf grey rgb dev.list extendrange
-#' @importFrom graphics abline box curve hist image layout legend lines par plot points polygon segments rect axis mtext
+#' @importFrom graphics abline box curve hist image layout legend lines par plot points polygon segments rect axis mtext plot
 #' @importFrom stats approx dbeta density dgamma dnorm dunif lm quantile rnorm weighted.mean coef
 #' @importFrom utils read.csv read.table write.table packageName txtProgressBar setTxtProgressBar
 #' @importFrom Rcpp evalCpp
@@ -16,69 +16,20 @@
 #' @name rbacon
 NULL
 
-# to be able to directly use copyCalibrationCurve, mix.curves, pMC.age & age.pMC
+# to be able to directly use copyCalibrationCurve, mix.curves, pMC.age & age.pMC (without having to type IntCal:: first)
 library(IntCal)
 
-# see if/where R subsamples .out, as the .out file is identical if seed is set!
+#if(!exists("info"))  info <- c() # a user reported that rbacon was looking for info but not finding it. Not sure if this is a good idea, so, commenting it
 
-# do: add.date() how indicate postbomb curve to use?
+# check if greyscale is still done OK, with only 1 region the darkest instead of all depths having the same darkest grey. Also see if cal blobs might need to be plotted wider as currently poorly visible. Some reports that greyscale looks bad when exporting from Rstudio.
 
-# rplum integration of the c++ code:
-# no differences between the files for ranfun.h, Matrix.h, Matrix.cpp or events.cpp
-# vector.cpp:  added lines 30 and 32 from rplum's
-# twalk.h: files are the same except for code to check and report seeds in rbacon's
-# ranfun.cpp: no differences except for lines commented in order to debug the seed problem
+# do: add.date() how indicate postbomb curve to use? 
 
-# input.h: added line 50 int plum; from rplum, added lines 97, 99, 101 from rplum
+# add explanation new mem prior 
 
-# input.cpp: 
-# line 8, BUFFZ 1024 in rbacon, but 50000 in rplum. Which one is best?
-# added lines 61-2, plum = 0; int more_pars = 0; from rplum
-# added lines 152-62 for the Plum 'calibration curve'. 
-# line 220, numofpars == 12 in rplum, 11 in rbacon. Seed. Check if this should be 12 for seed to work.
-# lines 233, 240, added more_pars from rplum
-# added lines 260-368, which define a new function OutputFiles
+# check if par(op) can be done better, as axis settings are off after agedepth()
 
-# cal.h: added class Plum (lines 1066-1174), added lines 1192-3 is_210Pb, rho, delta, added lines 1263-6 Is210Pb(), added lines 1327-30
-
-# bacon.h: 
-# line 36 LA_CONST added, line 77 added MaxYrTheta0Plum
-# line 73, added *x, changed x to X
-# line 90-112, added priorPhiU, priorPSU
-# lines 133 and onward, x replaced by X, added lines 
-# lines 223-8, copied from rplum.h
-# lines 269-71: x0 becomes X0
-# lines 181-5: added lines from rplum.h
-# added lines 185-188
-# lines 293-4: x became X
-# lines 301, 303, changed lines
-# line 312, x became X
-# changed line 318
-# added lines 342-400 for Plum
-# added lines 437-451 from Plum
-# commented lines 453-464 on hiatus... can we?
-# added lines 408-427 for Plum
-# commented lines 489-90 on int rt - SetThetas(x);
-# added delete x in line 422
-# changed x to X in lines 423, 424, 446, 447
-# what does prime do in function starting line 448?
-# added explanation lines from rplum in lines 449-450
-# replaced lines 459-476 with updated lines 479-517
-# changed last lines, 518 & 520 to return U
-
-# according to Marco, the 'fat' age-model bug was in bacon.h; for (int j=0; j<(m-1); j++) has to go until m, not m-1
-
-# bacon.cpp: 
-# added line 127 All.outputFiles(outputFile1)
-# note that lines 61-6 still uncommented, but are active in rplum
-
-#### since so many changes in bacon.h, simply copied rplum's bacon.h into rbacon
-
-# check with Andres:
-# input.cpp, line 8: BUFFZ 1024 in rbacon, but 50000 in rplum. Which one is best?
-# kernel.cpp, lines 151-9: added if(vector_cmp(x,xp,n) != 1) { ... else return -1.0; from rplum and commented it.
-
-# done: seed is working again
+# done: seed is working again. But, see if/where R subsamples .out, as the .out file is identical if seed is set!
 
 # for future versions: investigate the slowness of plotting after the Bacon run (not only dates, also the model's 95% ranges etc.), can ssize be predicted more accurately?, accrate.age.ghost is black all through - needs to have sections with lower maximum amount of grey, check fs::path(dir, data_name) as cross-platform alternative to specifying paths, why do we warn that "acc.shape cannot be equal to acc.mean"?, find a way to get rid of accrate.age.ghost's overly low accrates at core bottoms, check flux, add vignette(s), produce greyscale proxy graph with proxy uncertainties?, smooth bacon, check/adapt behaviour of AgesOfEvents around hiatuses, add function to estimate best thickness, F14C, if hiatus or boundary plot acc.posts of the individual sections?, allow for asymmetric cal BP errors (e.g. read from files), make more consistent use of dark for all functions (incl. flux and accrate.age.ghost), remove darkest?, introduce write.Bacon function to write files only once user agrees with the model, can we change from using files to using memory only?, proxy.ghost very slow with long/detailed cores - optimization possible?, check again if/how/when Bacon gets confused by Windows usernames with non-ascii characters (works fine on Mac)
 
@@ -92,7 +43,7 @@ library(IntCal)
 #' and through millions of Markov Chain Monte Carlo (MCMC) iterations estimates
 #' the accumulation rate (in years/cm; so more correctly, sedimentation times) for each of these sections.
 #' Combined with an estimated starting date for the first section, these accumulation rates then form the age-depth model.
-#' The accumulation rates are constrained by prior information on the accumulation rate (\code{acc.mean, acc.shape)} and its
+#' The accumulation rates are constrained by prior information on the accumulation rate (\code{acc.mean, acc.shape}) and its
 #' variability between neighbouring depths, or "memory" (\code{mem.mean, mem.strength}). Hiatuses can be introduced as well, also constrained by prior information (\code{hiatus.max}).
 #'
 #' Although Bacon works with any kind of absolute dates (e.g., OSL, tephra or other dates on a calendar scale),
@@ -103,9 +54,10 @@ library(IntCal)
 #' SH1-2 or SH3 for the southern hemisphere; Hua et al., 2013). See \url{http://calib.org/CALIBomb/} if you are unsure which
 #' postbomb curve you need. If Bacon finds postbomb dates (negative 14C ages) and you haven't specified a postbomb curve,
 #' you will be prompted. Provide postbomb curves as, e.g., \code{postbomb=1} for the NH1 postbomb curve (2 for NH2, 3 for NH3, 4 for SH1-2, 5 for SH3).
-#'
 #' For calendar dates, i.e. dates that are already on the calendar scale and thus should not be calibrated, set\code{cc=0}.
 #'
+#' From version 2.5.1 on (i.e., since February 2021), the default memory prior has changed to \code{mem.mean=0.5} and \code{mem.strength=10}. Previously used c++ code contained a bug which caused the prior information for the memory not to be taken into account correctly. Now that this bug has been repaired, the default memory prior has been updated such that it should work for most types of cores, and should result in similar output to previous versions of Bacon. There is no need to re-do previous runs. However, it is considered good practice to test the impact of different settings on a site's age-depth model (e.g., thick, acc.mean, acc.shape, mem.mean, acc.strength).
+
 #' @param core Name of the core, given using quotes. Defaults to one of the cores provided with rbacon, \code{core="MSB2K"}.
 #' An alternative core provided with this package is RLGH3 (Jones et al., 1989).
 #' To run your own core, produce a .csv file with the dates as outlined in the manual, add a folder with the core's name to the default directory for cores (see \code{coredir}), and save the .csv file there. For example, the file's location and name could be \code{Bacon_runs/MyCore/MyCore.csv}. Then run Bacon as follows: \code{Bacon("MyCore")}
@@ -132,11 +84,11 @@ library(IntCal)
 #' @param acc.mean The accumulation rate prior consists of a gamma distribution with two parameters. Its mean is set by acc.mean (default \code{acc.mean=20} yr/cm (or whatever age or depth units are chosen),
 #' which can be changed to, e.g., 5, 10 or 50 for different kinds of deposits). Multiple values can be given in case of hiatuses or boundaries, e.g., Bacon(hiatus.depths=23, acc.mean=c(5,20))
 #' @param mem.strength The prior for the memory (dependence of accumulation rate between neighbouring depths) is a beta distribution, which looks much like the gamma distribution.
-#'  but its values are always between 0 (no assumed memory) and 1 (100\% memory). Its default settings of \code{mem.strength=4}
-#'  (higher values result in more peaked shapes) allow for a large range of posterior memory values.
+#'  but its values are always between 0 (no assumed memory) and 1 (100\% memory). Its default settings of \code{mem.strength=10}
+#'  (higher values result in more peaked shapes) allow for a large range of posterior memory values. Please note that the default memory prior has been updated from rbacon version 2.5.1. on, to repair a bug. 
 #' @param mem.mean The prior for the memory is a beta distribution, which looks much like the gamma distribution but
-#' its values are always between 0 (no assumed memory) and 1 (100\% memory). Its default settings of \code{mem.mean=0.7}
-#' allow for a large range of posterior memory values.
+#' its values are always between 0 (no assumed memory) and 1 (100\% memory). Its default settings of \code{mem.mean=0.5}
+#' allow for a large range of posterior memory values. Please note that the default memory prior has been updated from rbacon version 2.5.1. on, to repair a bug. 
 #' @param boundary The assumed depths of any boundary, which divides sections of different accumulation rate regimes (e.g., as indicated by major change in the stratigraphy). No hiatus is assumed between these sections, and memory is reset crossing the boundary. Different accumulation priors can be set for the sections above and below the boundary, e.g., \code{acc.mean=c(5, 20)}. See also \code{hiatus.depths}, \code{mem.mean}, \code{acc.mean} and \code{acc.shape}. Setting many boundaries might not work, and having more than one boundary per model section (see \code{'thick'}) might not work either.
 #' @param hiatus.depths The assumed depths for any hiatus should be provided as, e.g.,
 #' \code{hiatus.depths=20} for one at 20cm depth, and \code{hiatus.depths=c(20,40)} for two hiatuses at 20 and 40 cm depth.
@@ -215,21 +167,21 @@ library(IntCal)
 #'
 #' Christen, J.A., Perez E., S., 2010. A new robust statistical model for radiocarbon data. Radiocarbon 51, 1047-1059. 
 #'
-#' Reimer et al., 2020. The IntCal20 Northern Hemisphere radiocarbon age calibration curve (0–55 cal kBP). Radiocarbon 62. doi: 10.1017/RDC.2020.41
+#' Reimer et al., 2020. The IntCal20 Northern Hemisphere radiocarbon age calibration curve (0–55 cal kBP). Radiocarbon 62. \doi{10.1017/RDC.2020.41}
 #'
-#' Hogg et al. 2020 SHCal20 Southern Hemisphere calibration, 0-55,000 years cal BP. Radiocarbon 62. doi: 10.1017/RDC.2020.59
+#' Hogg et al. 2020 SHCal20 Southern Hemisphere calibration, 0-55,000 years cal BP. Radiocarbon 62. \doi{10.1017/RDC.2020.59}
 #'
-#' Hughen et al. 2020 Marine20-the marine radiocarbon age calibration curve (0-55,000 cal BP). Radiocarbon 62. doi: 10.1017/RDC.2020.68.
+#' Hughen et al. 2020 Marine20-the marine radiocarbon age calibration curve (0-55,000 cal BP). Radiocarbon 62. \doi{10.1017/RDC.2020.68}
 #'
 #' Hua, Q., Barbetti, M., Rakowski, A.Z., 2013. Atmospheric radiocarbon for the period 1950-2010.
-#' Radiocarbon 55(4), <doi:10.2458/azu_js_rc.v55i2.16177>.
+#' Radiocarbon 55(4), \doi{10.2458/azu_js_rc.v55i2.16177}
 #'
 #' Jones, V.J., Stevenson, A.C., Battarbee, R.W., 1989. Acidification of lakes in Galloway, south west Scotland
 #' - a diatom and pollen study of the post-glacial history of the Round Loch of Glenhead.
 #' Journal of Ecology 77: 1-23.
 #'
 #' @export
-Bacon <- function(core="MSB2K", thick=5, coredir="", prob=0.95, d.min=NA, d.max=NA, add.bottom=TRUE, d.by=1, seed=NA, depths.file=FALSE, depths=c(), depth.unit="cm", age.unit="yr", unit=depth.unit, acc.shape=1.5, acc.mean=20, mem.strength=4, mem.mean=0.7, boundary=NA, hiatus.depths=NA, hiatus.max=10000, add=c(), after=.0001/thick, cc=1, cc1="IntCal20", cc2="Marine20", cc3="SHCal20", cc4="ConstCal", ccdir="", postbomb=0, delta.R=0, delta.STD=0, t.a=3, t.b=4, normal=FALSE, suggest=TRUE, accept.suggestions=FALSE, reswarn=c(10,200), remember=TRUE, ask=TRUE, run=TRUE, defaults="defaultBacon_settings.txt", sep=",", dec=".", runname="", slump=c(), remove=FALSE, BCAD=FALSE, ssize=2000, th0=c(), burnin=min(500, ssize), MinAge=c(), MaxAge=c(), MinYr=MinAge, MaxYr=MaxAge, cutoff=.1, plot.pdf=TRUE, dark=1, date.res=100, age.res=200, yr.res=age.res, close.connections=TRUE, verbose=TRUE, ...) {
+Bacon <- function(core="MSB2K", thick=5, coredir="", prob=0.95, d.min=NA, d.max=NA, add.bottom=TRUE, d.by=1, seed=NA, depths.file=FALSE, depths=c(), depth.unit="cm", age.unit="yr", unit=depth.unit, acc.shape=1.5, acc.mean=20, mem.strength=10, mem.mean=0.5, boundary=NA, hiatus.depths=NA, hiatus.max=10000, add=c(), after=.0001/thick, cc=1, cc1="IntCal20", cc2="Marine20", cc3="SHCal20", cc4="ConstCal", ccdir="", postbomb=0, delta.R=0, delta.STD=0, t.a=3, t.b=4, normal=FALSE, suggest=TRUE, accept.suggestions=FALSE, reswarn=c(10,200), remember=TRUE, ask=TRUE, run=TRUE, defaults="defaultBacon_settings.txt", sep=",", dec=".", runname="", slump=c(), remove=FALSE, BCAD=FALSE, ssize=2000, th0=c(), burnin=min(500, ssize), MinAge=c(), MaxAge=c(), MinYr=MinAge, MaxYr=MaxAge, cutoff=.01, plot.pdf=TRUE, dark=1, date.res=100, age.res=200, yr.res=age.res, close.connections=TRUE, verbose=TRUE, ...) {
   # Check coredir and if required, copy example file in core directory
   coredir <- assign_coredir(coredir, core, ask)
   if(core == "MSB2K" || core == "RLGH3") {
@@ -280,6 +232,10 @@ Bacon <- function(core="MSB2K", thick=5, coredir="", prob=0.95, d.min=NA, d.max=
       }
     }
 
+  # check values for the prior's mean, Jan 2021
+  if(mem.mean < 0 || mem.mean >1)
+    stop("The prior for the mean of the memory should be between 0 and 1", FALSE)
+    
   if(!is.na(boundary[1]))
     boundary <- sort(unique(boundary))
   if(!is.na(hiatus.depths[1])) {
@@ -369,13 +325,13 @@ Bacon <- function(core="MSB2K", thick=5, coredir="", prob=0.95, d.min=NA, d.max=
           sugg <- pretty(thick*(info$K/min(reswarn)), 10)
           sugg <- min(sugg[sugg>0])
           if(accept.suggestions) 
-            ans <- sugg else 
+            ans <- "y" else 
               ans <- readline(message(" Warning, the current value for thick, ", thick, ", will result in very few age-model sections (", info$K, ", not very flexible). Suggested maximum value for thick: ", sugg, " OK? (y/n) "))
         } else
           if(info$K > max(reswarn)) {
             sugg <- max(pretty(thick*(info$K/max(reswarn))))
             if(accept.suggestions) 
-              ans <- sugg else
+              ans <- "y" else
                 ans <- readline(message(" Warning, the current value for thick, ", thick, ", will result in very many age-model sections (", info$K, ", possibly hard to run). Suggested minimum value for thick: ", sugg, " OK? (y/n) "))
           }
     if(tolower(substr(ans, 1, 1)) == "y") {